node_backend/requirements.txt
flask
xgboost
gradio_client
flask-cors
numpy
scikit-learn
prophet
pytrends
python-dotenv
openai
transformers
torch
torchvision
numpy
Pillow
matplotlib
pandas
ftfy
regex
cli_interrogator
tqdm
clip @ git+https://github.com/openai/CLIP.git